requirements.txt
psycopg2-binary == 2.9.1
jinja2 == 3.0.3
pdfkit == 1.0.0

